// backoffice-frontend/src/app/core/interfaces/product.ts
export interface Product {

    comments: String;
    category: number;
    availability: Boolean;
    tig_id: number;
    price: number;
    price_on_sale: number;
    discount: number;
    sale: Boolean;
    owner: String;
    unit: String;
    name: String;
    quantityInStock: number;
    quantity_sold: number;
}